--- a/docs/src/manual/constraints.md
+++ b/docs/src/manual/constraints.md
@@ -165,7 +165,7 @@ DocTestSetup = quote
     @constraint(model, con, x <= 1);
     @objective(model, Max, -2x);
     optimize!(model);
-    mock = backend(model).optimizer.model;
+    mock = unsafe_backend(model);
     MOI.set(mock, MOI.TerminationStatus(), MOI.OPTIMAL)
     MOI.set(mock, MOI.DualStatus(), MOI.FEASIBLE_POINT)
     MOI.set(mock, MOI.ConstraintDual(), optimizer_index(con), -2.0)