--- a/test/generate_and_solve.jl
+++ b/test/generate_and_solve.jl
@@ -56,7 +56,7 @@ using JuMP
         )
         JuMP.optimize!(m)
 
-        mockoptimizer = JuMP.backend(m).optimizer.model
+        mockoptimizer = JuMP.unsafe_backend(m)
         MOI.set(mockoptimizer, MOI.TerminationStatus(), MOI.OPTIMAL)
         MOI.set(mockoptimizer, MOI.RawStatusString(), "solver specific string")
         MOI.set(mockoptimizer, MOI.ObjectiveValue(), -1.0)
@@ -204,7 +204,6 @@ using JuMP
                 MOIU.Model{Float64}(),
                 eval_objective_value = false,
             ),
-            caching_mode = MOIU.AUTOMATIC,
         )
         @variable(m, x == 1.0, Int)
         @variable(m, y, Bin)
@@ -233,7 +232,7 @@ using JuMP
 
         MOIU.attach_optimizer(m)
 
-        mockoptimizer = JuMP.backend(m).optimizer.model
+        mockoptimizer = JuMP.unsafe_backend(m)
         MOI.set(mockoptimizer, MOI.TerminationStatus(), MOI.OPTIMAL)
         MOI.set(mockoptimizer, MOI.RawStatusString(), "solver specific string")
         MOI.set(mockoptimizer, MOI.ObjectiveValue(), 1.0)
@@ -314,7 +313,7 @@ using JuMP
         )
         JuMP.optimize!(m)
 
-        mockoptimizer = JuMP.backend(m).optimizer.model
+        mockoptimizer = JuMP.unsafe_backend(m)
         MOI.set(mockoptimizer, MOI.TerminationStatus(), MOI.OPTIMAL)
         MOI.set(mockoptimizer, MOI.RawStatusString(), "solver specific string")
         MOI.set(mockoptimizer, MOI.ObjectiveValue(), -1.0)
@@ -627,7 +626,7 @@ c2: x + y <= 1.0
         )
         JuMP.optimize!(m)
 
-        mock = JuMP.backend(m).optimizer.model
+        mock = JuMP.unsafe_backend(m)
         MOI.set(mock, MOI.TerminationStatus(), MOI.OPTIMAL)
         MOI.set(mock, MOI.ResultCount(), 2)