--- a/test/model.jl
+++ b/test/model.jl
@@ -203,11 +203,13 @@ end
 
 function test_bridges_automatic()
     # optimizer not supporting Interval
-    model = Model(() -> MOIU.MockOptimizer(SimpleLPModel{Float64}()))
-    @test JuMP.bridge_constraints(model)
+    model = Model(
+        () -> MOIU.MockOptimizer(SimpleLPModel{Float64}());
+        bridge_constraints = true,
+    )
     @test JuMP.backend(model) isa MOIU.CachingOptimizer
     @test JuMP.backend(model).optimizer isa MOI.Bridges.LazyBridgeOptimizer
-    @test JuMP.backend(model).optimizer.model isa MOIU.MockOptimizer
+    @test JuMP.unsafe_backend(model) isa MOIU.MockOptimizer
     @variable model x
     cref = @constraint model 0 <= x + 1 <= 1
     @test cref isa JuMP.ConstraintRef{
@@ -228,12 +230,12 @@ function test_bridges_automatic_with_cache()
             SimpleLPModel{Float64}(),
             needs_allocate_load = true,
         ),
+        bridge_constraints = true,
     )
-    @test JuMP.bridge_constraints(model)
     @test JuMP.backend(model) isa MOIU.CachingOptimizer
     @test JuMP.backend(model).optimizer isa MOI.Bridges.LazyBridgeOptimizer
     @test JuMP.backend(model).optimizer.model isa MOIU.CachingOptimizer
-    @test JuMP.backend(model).optimizer.model.optimizer isa MOIU.MockOptimizer
+    @test JuMP.unsafe_backend(model) isa MOIU.MockOptimizer
     @variable model x
     err = ErrorException(
         "There is no `optimizer_index` as the optimizer is not " *
@@ -258,27 +260,10 @@ function test_bridges_automatic_with_cache()
     @test_throws err optimizer_index(cref)
 end
 
-function test_bridges_automatic_disabled()
-    # Automatic bridging disabled with `bridge_constraints` keyword
-    model = Model(
-        () -> MOIU.MockOptimizer(SimpleLPModel{Float64}()),
-        bridge_constraints = false,
-    )
-    @test !JuMP.bridge_constraints(model)
-    @test JuMP.backend(model) isa MOIU.CachingOptimizer
-    @test !(JuMP.backend(model).optimizer isa MOI.Bridges.LazyBridgeOptimizer)
-    @variable model x
-    err = ErrorException(
-        "Constraints of type MathOptInterface.ScalarAffineFunction{Float64}-in-MathOptInterface.Interval{Float64} are not supported by the solver, try using `bridge_constraints=true` in the `JuMP.Model` constructor if you believe the constraint can be reformulated to constraints supported by the solver.",
-    )
-    @test_throws err @constraint model 0 <= x + 1 <= 1
-end
-
 function test_bridges_direct()
     # No bridge automatically added in Direct mode
     optimizer = MOIU.MockOptimizer(SimpleLPModel{Float64}())
     model = JuMP.direct_model(optimizer)
-    @test !JuMP.bridge_constraints(model)
     @variable model x
     err = ErrorException(
         "Constraints of type MathOptInterface.ScalarAffineFunction{Float64}-in-MathOptInterface.Interval{Float64} are not supported by the solver.",
@@ -303,7 +288,7 @@ function mock_factory()
 end
 
 function test_bridges_add_before_con_model_optimizer()
-    model = Model(mock_factory)
+    model = Model(mock_factory; bridge_constraints = true)
     @variable(model, x)
     JuMP.add_bridge(model, NonnegativeBridge)
     c = @constraint(model, x in Nonnegative())
@@ -318,7 +303,7 @@ function test_bridges_add_before_con_set_optimizer()
     @variable(model, x)
     c = @constraint(model, x in Nonnegative())
     JuMP.add_bridge(model, NonnegativeBridge)
-    set_optimizer(model, mock_factory)
+    set_optimizer(model, mock_factory; bridge_constraints = true)
     JuMP.optimize!(model)
     @test 1.0 == @inferred JuMP.value(x)
     @test 1.0 == @inferred JuMP.value(c)
@@ -326,7 +311,7 @@ function test_bridges_add_before_con_set_optimizer()
 end
 
 function test_bridges_add_after_con_model_optimizer()
-    model = Model(mock_factory)
+    model = Model(mock_factory; bridge_constraints = true)
     @variable(model, x)
     flag = true
     try
@@ -381,7 +366,7 @@ function test_bridges_add_bridgeable_con_set_optimizer()
     constraint = ScalarConstraint(x, Nonnegative())
     bc = BridgeableConstraint(constraint, NonnegativeBridge)
     c = add_constraint(model, bc)
-    set_optimizer(model, mock_factory)
+    set_optimizer(model, mock_factory; bridge_constraints = true)
     JuMP.optimize!(model)
     @test 1.0 == @inferred JuMP.value(x)
     @test 1.0 == @inferred JuMP.value(c)
@@ -389,19 +374,13 @@ function test_bridges_add_bridgeable_con_set_optimizer()
 end
 
 function test_bridge_graph_false()
-    model = Model(mock_factory, bridge_constraints = false)
+    model = Model(mock_factory)
     @variable(model, x)
     @test_throws(
         ErrorException(
-            "Cannot add bridge if `bridge_constraints` was set to `false` in " *
-            "the `Model` constructor.",
-        ),
-        add_bridge(model, NonnegativeBridge)
-    )
-    @test_throws(
-        ErrorException(
-            "Cannot print bridge graph if `bridge_constraints` was set to " *
-            "`false` in the `Model` constructor.",
+            "`In order to print the bridge graph, you must pass " *
+            "`bridge_constraints=true` to `Model`, i.e., " *
+            "`Model(optimizer; bridge_constraints = true)`.",
         ),
         print_bridge_graph(model)
     )
@@ -509,7 +488,7 @@ end
 function dummy_optimizer_hook(::JuMP.AbstractModel) end
 
 function copy_model_style_mode(use_copy_model, caching_mode, filter_mode)
-    model = Model(caching_mode = caching_mode)
+    model = Model()
     model.optimize_hook = dummy_optimizer_hook
     data = DummyExtensionData(model)
     model.ext[:dummy] = data
@@ -595,12 +574,6 @@ end
 
 function test_copy_model_base_auto()
     return copy_model_style_mode(false, MOIU.AUTOMATIC, false)
-end
-function test_copy_model_jump_manual()
-    return copy_model_style_mode(true, MOIU.MANUAL, false)
-end
-function test_copy_model_base_manual()
-    return copy_model_style_mode(false, MOIU.MANUAL, false)
 end
 
 function test_copy_direct_mode()
@@ -782,7 +755,7 @@ function test_copy_conflict()
     )
     JuMP.optimize!(model)
 
-    mockoptimizer = JuMP.backend(model).optimizer.model
+    mockoptimizer = JuMP.unsafe_backend(model)
     MOI.set(mockoptimizer, MOI.TerminationStatus(), MOI.INFEASIBLE)
     MOI.set(mockoptimizer, MOI.ConflictStatus(), MOI.CONFLICT_FOUND)
     MOI.set(