This is just a proof-of-concept to look at what changes we can make to JuMP to address the "time-to-first-solve" issue (https://github.com/jump-dev/MathOptInterface.jl/issues/1313). It is essentially a repeat of https://github.com/jump-dev/MathOptInterface.jl/pull/1252, but at the JuMP level instead of MOI. 

## Background 

The main issues are:
 * Bridges are good, and have a small run-time cost if un-used
 * Due to the design, bridges are not type stable. The main reasons are:
   * New bridges can be added by the user
   * Variable bridges can return a function of a variable, instead of just a variable
   * Determining whether a variable is bridged requires a run-time lookup.
 * Because of the type instability, bridges don't precompile well, and they have a large compile-time cost.
 * This compile time-cost happens even if the bridges are not used. As one example:
   * To set the objective, we must first delete the current objective. Because we can only tell at runtime if the current objective is bridged, inference starts exploring the branch that deals with the bridges. Objective bridges lead to the graph, which leads to other bridges, and so inference continues until we hit the type-instabilities and it gives up.
 * One attempt at solving this was https://github.com/jump-dev/MathOptInterface.jl/pull/1252. But this tried to squeeze the logic into CachingOptimizer, and tied CachingOptimizer and LazyBridgeOptimizer together too much.
 * A related PR is https://github.com/jump-dev/JuMP.jl/pull/2520. If JuMP's backend is not a concrete type, we don't need the CachingOptimizer to also be non-concrete.

## Demo

By default, the `LazyBridgeOptimizer` is omitted. But, if you do something that requires bridges, they are added:
```Julia
julia> model = Model(Clp.Optimizer)
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: EMPTY_OPTIMIZER
Solver name: Clp

julia> backend(model)
MOIU.CachingOptimizer{Clp.Optimizer, MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}}
in state EMPTY_OPTIMIZER
in mode AUTOMATIC
with model cache MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}
  fallback for MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}
with optimizer Clp.Optimizer

julia> @variable(model, x[1:4] in MOI.Nonnegatives(4))
4-element Vector{VariableRef}:
 x[1]
 x[2]
 x[3]
 x[4]

julia> backend(model)
MOIU.CachingOptimizer{MOIB.LazyBridgeOptimizer{MOIU.CachingOptimizer{Clp.Optimizer, MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}}}, MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}}
in state EMPTY_OPTIMIZER
in mode AUTOMATIC
with model cache MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}
  fallback for MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}
with optimizer MOIB.LazyBridgeOptimizer{MOIU.CachingOptimizer{Clp.Optimizer, MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}}}
  with 0 variable bridges
  with 0 constraint bridges
  with 0 objective bridges
  with inner model MOIU.CachingOptimizer{Clp.Optimizer, MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}}
    in state ATTACHED_OPTIMIZER
    in mode AUTOMATIC
    with model cache MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}
      fallback for MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}
    with optimizer Clp.Optimizer
```

You can also pass `bridge_constraints = true` to force-add the bridges:
```julia
julia> model = Model(Clp.Optimizer; bridge_constraints = true)
A JuMP Model
Feasibility problem with:
Variables: 0
Model mode: AUTOMATIC
CachingOptimizer state: EMPTY_OPTIMIZER
Solver name: Clp

julia> backend(model)
MOIU.CachingOptimizer{MOIB.LazyBridgeOptimizer{MOIU.CachingOptimizer{Clp.Optimizer, MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}}}, MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}}
in state EMPTY_OPTIMIZER
in mode AUTOMATIC
with model cache MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}
  fallback for MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}
with optimizer MOIB.LazyBridgeOptimizer{MOIU.CachingOptimizer{Clp.Optimizer, MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}}}
  with 0 variable bridges
  with 0 constraint bridges
  with 0 objective bridges
  with inner model MOIU.CachingOptimizer{Clp.Optimizer, MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}}
    in state ATTACHED_OPTIMIZER
    in mode AUTOMATIC
    with model cache MOIU.UniversalFallback{MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}}
      fallback for MOIU.GenericModel{Float64, MOIU.ModelFunctionConstraints{Float64}}
    with optimizer Clp.Optimizer
```

## Alternative

The alternative is to change the default of `bridge_constraints` to `false`, and also fix the double abstract-type issue. 

This would give everyone a speed-up if they don't use bridges, and for models which use bridges, they can just go
```Julia
Model(GLPK.Optimizer, bridge_constraints = true)
```
This would have the added benefit of giving people greater insight into when bridges get used.

## Benchmark

```Julia
using JuMP, Clp

### No bridging

function foo()
    model = Model(Clp.Optimizer)
    set_silent(model)
    @variable(model, x >= 0)
    @constraint(model, 2x + 1 <= 1)
    @objective(model, Max, 1.0 * x)
    optimize!(model)
end

### ObjectiveFunction bridging

function foo_obj()
    model = Model(Clp.Optimizer)
    set_silent(model)
    @variable(model, x >= 0)
    @constraint(model, 2x + 1 <= 1)
    @objective(model, Max, x)
    optimize!(model)
end

### variable bridging

function foo_var()
    model = Model(Clp.Optimizer)
    set_silent(model)
    @variable(model, x[1:1] in MOI.Nonnegatives(1))
    @constraint(model, 2x[1] + 1 <= 1)
    @objective(model, Max, 1.0 * x[1])
    optimize!(model)
end

### constraint bridging

function foo_con()
    model = Model(Clp.Optimizer)
    set_silent(model)
    @variable(model, x[1:2])
    @constraint(model, x in MOI.Nonpositives(2))
    @objective(model, Max, 1.0 * x[1])
    optimize!(model)
end

if get(ARGS, 1, "") == "--var"
    @time foo_var()
    @time foo_var()
elseif get(ARGS, 1, "") == "--con"
    @time foo_con()
    @time foo_con()
elseif get(ARGS, 1, "") == "--obj"
    @time foo_obj()
    @time foo_obj()
else
    @time foo()
    @time foo()
end
```

Current `master`
```
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Documents/JuMP/performance/auto-cache/new_auto.jl
 15.042563 seconds (19.82 M allocations: 1.149 GiB, 2.70% gc time, 36.00% compilation time)
  0.001562 seconds (4.42 k allocations: 327.695 KiB)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Documents/JuMP/performance/auto-cache/new_auto.jl --var
 17.091837 seconds (21.71 M allocations: 1.264 GiB, 2.83% gc time, 33.98% compilation time)
  0.001261 seconds (4.74 k allocations: 365.414 KiB)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Documents/JuMP/performance/auto-cache/new_auto.jl --con
 15.977032 seconds (20.92 M allocations: 1.219 GiB, 2.91% gc time, 39.28% compilation time)
  0.001164 seconds (4.42 k allocations: 325.242 KiB)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Documents/JuMP/performance/auto-cache/new_auto.jl --obj
 15.682321 seconds (20.06 M allocations: 1.164 GiB, 3.10% gc time, 35.60% compilation time)
  0.001150 seconds (4.43 k allocations: 328.055 KiB)
```

This PR

```
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Documents/JuMP/performance/auto-cache/new_auto.jl 
  7.989580 seconds (4.24 M allocations: 250.037 MiB, 0.85% gc time, 99.96% compilation time)
  0.000485 seconds (1.87 k allocations: 154.148 KiB)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Documents/JuMP/performance/auto-cache/new_auto.jl --var
 16.764737 seconds (23.05 M allocations: 1.349 GiB, 2.82% gc time, 29.02% compilation time)
  0.001286 seconds (4.88 k allocations: 377.680 KiB)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Documents/JuMP/performance/auto-cache/new_auto.jl --con
 15.583102 seconds (22.05 M allocations: 1.290 GiB, 3.06% gc time, 31.92% compilation time)
  0.001189 seconds (4.53 k allocations: 334.664 KiB)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Documents/JuMP/performance/auto-cache/new_auto.jl --obj
 14.417599 seconds (19.37 M allocations: 1.130 GiB, 3.01% gc time, 38.62% compilation time)
  0.001061 seconds (4.56 k allocations: 339.273 KiB)
```

The big change is really that its 2x as fast when bridges are not needed. Note that this uses MOI 0.9.21, so it doesn't account for all the recent changes I've been making. We should re-do this benchmark after MOI is released.

## TODO

TODOs before merging
 - [ ] Update to MOI 0.10
 - [ ] Better tests
 - [ ] Update documentation