@blegat 2021-05-21
`bridge_constraint` does not have the same effect as before as setting it to `false` won't prevent bridges to be automatically added if needed.
If were going to be breaking, we might as well change the name since it should be changed before v1.0 (`bridge_constraint` does not make sense given that we also bridge variables and objectives).
It should be `bridge_mode` with
* `AUTO_BRIDGING`: add if needed
* `ADD_BRIDGING`: add from the start
* `NO_BRIDGING`: don't add it even if needed