@odow 2021-05-28
Docs are failing because of https://github.com/jump-dev/MathOptInterface.jl/pull/1329. This shows that we're copying straight to GLPK, which should give a nice speedup though!