@odow 2021-09-27
Good news on the GLPK front, we're < 4 seconds, compared to 15 seconds in May.
```
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Desktop/bench.jl glpk --bridge
  6.649749 seconds (15.36 M allocations: 933.000 MiB, 4.98% gc time, 21.35% compilation time)
  0.000296 seconds (926 allocations: 47.953 KiB)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Desktop/bench.jl glpk --no-bridge
  3.869593 seconds (5.49 M allocations: 324.722 MiB, 2.53% gc time, 99.96% compilation time)
  0.000163 seconds (376 allocations: 33.266 KiB)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Desktop/bench.jl glpk --var      
  7.808467 seconds (18.06 M allocations: 1.094 GiB, 5.11% gc time, 15.47% compilation time)
  0.000547 seconds (1.20 k allocations: 66.953 KiB)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Desktop/bench.jl glpk --con
  8.049347 seconds (19.54 M allocations: 1.166 GiB, 4.39% gc time, 20.44% compilation time)
  0.000466 seconds (1.10 k allocations: 60.031 KiB)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. ~/Desktop/bench.jl glpk --obj
  6.487435 seconds (15.40 M allocations: 934.871 MiB, 4.42% gc time, 36.21% compilation time)
  0.000282 seconds (957 allocations: 50.484 KiB)
```