@odow 2021-10-03
I need to rerun the benchmark because I don't think I precompiled things properly, but this is the sort of thing our new benchmarking scripts can produce:

 * `time_A` is the current `master`
 * `time_B` is this PR
 * `time_ratio` is `time_B / time_A` (smaller is better)
  
```
julia> DataFrames.sort(summary, :time_ratio)
43×4 DataFrame
 Row │ models                             time_A     time_B    time_ratio 
     │ String                             Float64    Float64   Float64    
─────┼────────────────────────────────────────────────────────────────────
   1 │ logistic_regression.jl             103.054    28.8979     0.280414
   2 │ facility_location.jl                58.4608   23.1348     0.395732
   3 │ factory_schedule.jl                 30.1585   14.776      0.489946
   4 │ working_with_data_files.jl          64.0063   34.7838     0.543443
   5 │ getting_started_with_JuMP.jl        19.4115   12.3698     0.637238
   6 │ transp.jl                           16.0639   11.8066     0.734978
   7 │ urban_plan.jl                       18.4748   13.7019     0.741653
   8 │ diet.jl                             16.3324   12.232      0.748942
   9 │ cannery.jl                          16.4898   12.4013     0.752059
  10 │ n-queens.jl                         14.5118   11.0387     0.760667
  11 │ clnlbeam.jl                         30.106    22.9125     0.761061
  12 │ sudoku.jl                           14.5466   11.1206     0.76448
  13 │ multi.jl                            15.8926   12.1848     0.766693
  14 │ knapsack.jl                         14.1738   11.0646     0.780637
  15 │ network_flows.jl                    16.4801   13.4379     0.8154
  16 │ steelT3.jl                          19.493    15.9266     0.817045
  17 │ prod.jl                             19.1153   15.6991     0.821288
  18 │ mle.jl                              25.3167   20.7961     0.821437
  19 │ solvers_and_solutions.jl            14.2674   11.7488     0.823468
  20 │ portfolio.jl                        25.9414   21.4679     0.827553
  21 │ cutting_stock_column_generation.…   18.4845   15.5869     0.84324
  22 │ qcp.jl                              26.0444   22.0733     0.847525
  23 │ nlp_tricks.jl                       25.4686   21.6728     0.850962
  24 │ space_shuttle_reentry_trajectory…   48.6241   41.6138     0.855827
  25 │ power_systems.jl                    19.4688   16.7225     0.858938
  26 │ variables_constraints_objective.…   20.3478   17.5759     0.863775
  27 │ geographic_clustering.jl            27.6831   24.0654     0.869318
  28 │ rocket_control.jl                   39.3817   34.408      0.873706
  29 │ rosenbrock.jl                       23.0309   20.6918     0.898437
  30 │ performance_tips.jl                  7.94994   7.2595     0.913152
  31 │ tips_and_tricks.jl                  13.0748   12.3481     0.944421
  32 │ robust_uncertainty.jl               29.0863   27.7542     0.954202
  33 │ cluster.jl                          27.2482   26.5287     0.973594
  34 │ an_introduction_to_julia.jl          7.66265   7.48934    0.977383
  35 │ callbacks.jl                        17.4636   17.0972     0.979016
  36 │ benders_lazy_constraints.jl         19.3079   18.9655     0.98227
  37 │ finance.jl                          16.6231   16.3396     0.982942
  38 │ benders_decomposition.jl            18.8689   18.7684     0.994677
  39 │ min_ellipse.jl                      25.6519   25.6155     0.998581
  40 │ min_distortion.jl                   27.3975   27.4592     1.00225
  41 │ max_cut_sdp.jl                      26.702    28.3136     1.06035
  42 │ corr_sdp.jl                         25.4002   27.188      1.07039
  43 │ experiment_design.jl                28.3194   31.2174     1.10233
```