@odow 2021-10-14
I'm having second thoughts about this PR. Perhaps we just need to:

 * Fix the double-abstract-type issue (#2520)
 * Make the documentation much clearer and explicit that people should try `bridge_constraints = false` to get a speed boost
 * Rename `bridge_constraints` to `add_bridges`

`Model(GLPK.Optimizer; add_bridges = false)` isn't that bad. It's explicit in what it is doing, and it doesn't rely on this hacky trying to catch errors mechanism.